Add readme, delete old stuff
